Added a lifecycle method to compare previous and current props
We have to figure out the way to animiate actions like undo, redo
etc. To do this, we have to be able to compare previous and current
props and based on some differences aka the length of some arrays,
we can determine if user actually clicked undo, redo or whathever.